+++ packages/lumx-vue/src/components/badge/Stories/BadgeDefault.vue
@@ -1,0 +1,9 @@
+<template>
+    <Badge v-bind="$attrs">
+        <slot />
+    </Badge>
+</template>
+
+<script setup lang="ts">
+import { Badge } from '@lumx/vue';
+</script>