+++ packages/lumx-vue/src/components/badge/index.ts
@@ -1,0 +1,5 @@
+import Badge, { BadgeProps } from './Badge';
+import BadgeWrapper, { BadgeWrapperProps } from './BadgeWrapper';
+
+export { Badge, BadgeWrapper };
+export type { BadgeProps, BadgeWrapperProps };