+++ packages/site-demo/content/product/components/badge/vue/demo1.vue
@@ -1,0 +1,9 @@
+<template>
+    <Badge color="red">
+        <span>4</span>
+    </Badge>
+</template>
+
+<script setup lang="ts">
+import { Badge } from '@lumx/vue';
+</script>