+++ packages/site-demo/content/product/components/badge/vue/demo2.vue
@@ -1,0 +1,14 @@
+<template>
+    <Badge color="green">
+        <Icon :icon="mdiCheck" />
+    </Badge>
+
+    <Badge color="red">
+        <Icon :icon="mdiClose" />
+    </Badge>
+</template>
+
+<script setup lang="ts">
+import { mdiCheck, mdiClose } from '@lumx/icons';
+import { Badge, Icon } from '@lumx/vue';
+</script>