+++ packages/site-demo/content/product/components/badge/vue/demo3.vue
@@ -1,0 +1,9 @@
+<template>
+    <Badge>
+        <img src="https://i.pravatar.cc/128?img=32" alt="Logo" style="width: 16px; height: 16px" />
+    </Badge>
+</template>
+
+<script setup lang="ts">
+import { Badge } from '@lumx/vue';
+</script>